+++ domain/src/main/java/com/golandcoinc/domain/entities/data/StatsData.kt
@@ -1,7 +1,7 @@
 package com.golandcoinc.domain.entities.data
 
 data class StatsData(
-    val medianSpeedForTravel: Double,
+    val medianSpeedForTravel: Int,
     val totalDistanceTraveled: Double,
     val fuelVolume: Double,
     val fuelConsumption: Double
